+++ unpacked/_next/static/chunks/webpack.js
@@ -205,7 +205,7 @@
       );
     }),
     (b.miniCssF = function (e) {
-      return "static/css/8f2a2da44b8da8d5.css";
+      return "static/css/1033ddd3b7950a2f.css";
     }),
     (b.g = (function () {
       if ("object" == typeof globalThis) return globalThis;
@@ -357,4 +357,4 @@
     (d.push = o.bind(null, d.push.bind(d))),
     (b.nc = void 0);
 })();
-//# sourceMappingURL=webpack-c82c25c608202a42.js.map
+//# sourceMappingURL=webpack-e863b3dfca2be08d.js.map